requirements.txt
torch
torchvision
numpy
pillow
matplotlib
seaborn
tqdm
highlight-text
scikit-image
